requirements.txt
Pillow>=10.0.0
python-telegram-bot>=20.0
pyyaml>=6.0
numpy>=1.24.0
python-dotenv>=1.0.0
loguru>=0.7.0
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0  # Для асинхронной работы с PostgreSQL
aiohttp>=3.9.0  # Для асинхронных HTTP-запросов
